fix(data): handle cases where API returns single result instead of array
For some reason the API sends back a vare object rather than an array with a single object in it for instances where there is only a lone result. This causes the plugin to fail as it tries to loop over the object as if it were an array. Handle this case ratehr non-gracefully with a guard. We'll probably need to fix this in the webservices packages at some point...